requirements.txt
# Core dependencies
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Tracing
langfuse>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

